# api-sigea/openapi/sigec.yaml
openapi: 3.0.0
info:
  title: SigecAPIDMS
  description: Rest Api che consente la gestione e il monitoraggio degli eventi della regione Puglia
  version: v1
servers:
  - url: "http://localhost:9292/SigecAPI"
    description: LOCALE
  - url: ND
    description: SVILUPPO
  - url: ND
    description: COLLAUDO
  - url: ND
    description: PRODUZIONE
paths:
...
  "/eventi/{id}":
    get:
      tags:
        - Endpoint accessibili solo da SIGEA
      summary: API per la lettura di un evento
      parameters:
        - name: id
          in: path
          description: "Id dell'evento su cui effettuare la ricerca"
          required: true
          schema:
            type: number
            format: int64
      operationId: getEvento
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventoModel"
          description: Oggetto contenente le specifiche
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "409":
          $ref: "#/components/responses/409Conflict"
        "503":
          $ref: "#/components/responses/503ServiceUnavailable"
        default:
          $ref: "#/components/responses/default"
      security:
        - bearerAuth: []
      x-auth-type: "Application & Application User"
      x-throttling-tier: Unlimited
    delete:
      tags:
        - Endpoint accessibili solo da SIGEA
      summary: API per la cancellazione di un evento
      parameters:
        - name: id
          in: path
          description: "Id dell'evento da eliminare"
          required: true
          schema:
            type: number
            format: int64
      operationId: deleteEvento
      responses:
        "200":
          description: Cancellazione effettuata
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "409":
          $ref: "#/components/responses/409Conflict"
        "503":
          $ref: "#/components/responses/503ServiceUnavailable"
        default:
          $ref: "#/components/responses/default"
      security:
        - bearerAuth: []
      x-auth-type: "Application & Application User"
      x-throttling-tier: Unlimited
...
  /eventi/stati:
    get:
      tags:
        - Endpoint accessibili solo da SIGEA
      summary: API per la lettura degli stati
      operationId: getStatiEvento
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatoModel"
          description: Oggetto contenente le specifiche
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "503":
          $ref: "#/components/responses/503ServiceUnavailable"
        default:
          $ref: "#/components/responses/default"
      security:
        - bearerAuth: []
      x-auth-type: "Application & Application User"
      x-throttling-tier: Unlimited
...
  /utenti:
    put:
      tags:
        - Endpoint accessibili solo da SIGEA
      summary: "API per la profilazione dell'utente, i dati vengono estratti direttamente dal token"
      operationId: saveUtente
      parameters:
        - name: tipologia
          in: query
          description: "id tipologia da associare all'utente"
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Profilazione avvenuta correttamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DettaglioUtenteModel"
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "503":
          $ref: "#/components/responses/503ServiceUnavailable"
        default:
          $ref: "#/components/responses/default"
      security:
        - bearerAuth: []
      x-auth-type: "Application & Application User"
      x-throttling-tier: Unlimited
...
  /redazioni:
    get:
      tags:
        - UtenteAPI
      summary: API per la lettura delle redazioni
      operationId: getRedazioni
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RedazioneModel"
          description: Oggetto contenente le specifiche
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "503":
          $ref: "#/components/responses/503ServiceUnavailable"
        default:
          $ref: "#/components/responses/default"
      security:
        - bearerAuth: []
      x-auth-type: "Application & Application User"
      x-throttling-tier: Unlimited
...
  /mezzi:
    get:
      tags:
        - OntologiaAPI
      summary: API per la lettura dei mezzi
      operationId: getMezzi
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MezzoModel"
          description: Oggetto contenente le specifiche
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "503":
          $ref: "#/components/responses/503ServiceUnavailable"
        default:
          $ref: "#/components/responses/default"
      security:
        - bearerAuth: []
      x-auth-type: "Application & Application User"
      x-throttling-tier: Unlimited
  /prodotti:
    get:
      tags:
        - OntologiaAPI
      summary: API per la lettura dei prodotti
      operationId: getProdotti
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProdottoModel"
          description: Oggetto contenente le specifiche
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "503":
          $ref: "#/components/responses/503ServiceUnavailable"
        default:
          $ref: "#/components/responses/default"
      security:
        - bearerAuth: []
      x-auth-type: "Application & Application User"
      x-throttling-tier: Unlimited
  /tipologiemibact:
    get:
      tags:
        - OntologiaAPI
      summary: API per la lettura delle tipologie mibact
      operationId: getTipologieMIBACT
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TipologiaMIBACTModel"
          description: Oggetto contenente le specifiche
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "503":
          $ref: "#/components/responses/503ServiceUnavailable"
        default:
          $ref: "#/components/responses/default"
      security:
        - bearerAuth: []
      x-auth-type: "Application & Application User"
      x-throttling-tier: Unlimited
  /tipologieattivita:
    get:
      tags:
        - OntologiaAPI
      summary: API per la lettura delle tipologie delle attività
      operationId: getTipologieAttivita
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TipologiaAttivitaModel"
          description: Oggetto contenente le specifiche
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "503":
          $ref: "#/components/responses/503ServiceUnavailable"
        default:
          $ref: "#/components/responses/default"
      security:
        - bearerAuth: []
      x-auth-type: "Application & Application User"
      x-throttling-tier: Unlimited
...
  /valoriattrattivita:
    get:
      tags:
        - OntologiaAPI
      summary: API per la lettura dei valori di attrattività
      operationId: getValoriAttrattivita
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValoreAttrattivitaTuristicaModel"
          description: Oggetto contenente le specifiche
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "503":
          $ref: "#/components/responses/503ServiceUnavailable"
        default:
          $ref: "#/components/responses/default"
      security:
        - bearerAuth: []
      x-auth-type: "Application & Application User"
      x-throttling-tier: Unlimited
components:
  schemas:
...
    EventoModel:
      type: object
      properties:
        eventoId:
          type: number
          format: int64
          description: "Id dell'evento"
        tipo:
          type: string
          description: Tipologia (EVENTO/RASSEGNA)
        idMIBACT:
          type: string
          description: Id MIBACT
        tipologiaMIBACT:
          type: string
          description: "Tipologia MIBACT (Altro, Apertura straordinaria, Attività didattica, Concerto, Convegno/Conferenza, Festival, Inaugurazione, Incontro/Presentazione, Manifestazione fieristica, Manifestazione sportiva, Mostra, Nuova apertura, Proiezione, Seminario, Spettacolo di danza, Spettacolo teatrale, Visita guidata, Visita libera)"
        grandeEvento:
          type: boolean
          description: boolean che indica se si tratta di un grande evento secondo il MIBACT
        statoId:
          type: string
          description: Id dello stato corrente
        ultimeNote:
          type: string
          description: "Nota associata all'ultimo cambio di stato"
        descrizioneStato:
          type: string
          description: Descrizione dello stato corrente
        fgValidazionePregressa:
          type: boolean
          description: "Flag che indica se l'evento ha già subito in passato una validazione"
        dataIns:
          type: string
          format: date-time
          description: Data creazione
        progettoId:
          type: number
          format: int64
          description: Id del progetto associato
        titoloProgetto:
          type: string
          description: Titolo del progetto associato
        ticket:
          $ref: "#/components/schemas/TicketModel"
        ownerId:
          type: number
          format: int64
          description: "Id del creatore dell'evento"
        nomeOwner:
          type: string
          description: "Nome del creatore dell'evento"
        cognomeOwner:
          type: string
          description: "Cognome del creatore dell'evento"
        usernameOwner:
          type: string
          description: "Username del creatore dell'evento"
        emailOwner:
          type: string
          description: "Email del creatore dell'evento"
        telOwner:
          type: string
          description: "Telefono del creatore dell'evento"
        celOwner:
          type: string
          description: "Cellulare del creatore dell'evento"
        ownerIdUltimaModifica:
          type: number
          format: int64
          description: "Id dell'utente che ha effettuato l'ultima modifica"
        nomeOwnerUltimaModifica:
          type: string
          description: "Nome dell'utente che ha effettuato l'ultima modifica"
        cognomeOwnerUltimaModifica:
          type: string
          description: "Cognome dell'utente che ha effettuato l'ultima modifica"
        usernameOwnerUltimaModifica:
          type: string
          description: "Username dell'utente che ha effettuato l'ultima modifica"
        emailOwnerUltimaModifica:
          type: string
          description: "Email dell'utente che ha effettuato l'ultima modifica"
        telOwnerUltimaModifica:
          type: string
          description: "Telefono dell'utente che ha effettuato l'ultima modifica"
        celOwnerUltimaModifica:
          type: string
          description: "Cellulare dell'utente che ha effettuato l'ultima modifica"
        attivita:
          $ref: "#/components/schemas/AttivitaModel"
        richiestaAttivita:
          $ref: "#/components/schemas/RichiestaAttivitaModel"
        datiGenerali:
          $ref: "#/components/schemas/DatiGeneraliModel"
        accessibilita:
          $ref: "#/components/schemas/AccessibilitaModel"
        emailSet:
          type: array
          description: "Array di e-mail associate all'evento"
          items:
            $ref: "#/components/schemas/EmailModel"
        telefonoSet:
          type: array
          description: "Array di telefoni associati all'evento"
          items:
            $ref: "#/components/schemas/TelefonoModel"
        sitoSet:
          type: array
          description: "Array di siti web associati all'evento"
          items:
            $ref: "#/components/schemas/SitoModel"
        locationSet:
          type: array
          description: "Array di location associate all'evento"
          items:
            $ref: "#/components/schemas/LocationModel"
        immagineSet:
          type: array
          description: "Array di immagini associate all'evento"
          items:
            $ref: "#/components/schemas/ImmagineModel"
        documentoSet:
          type: array
          description: "Array di documenti associati all'evento"
          items:
            $ref: "#/components/schemas/DocumentoEventoModel"
        linkSet:
          type: array
          description: "Array di link streaming associati all'evento"
          items:
            $ref: "#/components/schemas/LinkModel"
        relazioneSet:
          type: array
          description: "Array di relazioni associate all'evento corrente"
          items:
            $ref: "#/components/schemas/RelazioneModel"
        logSet:
          type: array
          description: "Array di log associati all'evento"
          items:
            $ref: "#/components/schemas/LogEventoModel"
        periodoSet:
          type: array
          description: "Array di periodi associati all'evento"
          items:
            $ref: "#/components/schemas/PeriodoModel"
...
    MezzoModel:
      type: object
      properties:
        mezzoId:
          type: string
          description: Id del mezzo
        mezzo:
          type: string
          description: Descrizione del mezzo
...
    StatoModel:
      type: object
      properties:
        statoId:
          type: string
          description: Id dello stato
        descrizione:
          type: string
          description: Descrizione dello stato
        statoInizio:
          type: boolean
          description: Indica se è lo stato di partenza del workflow
        statoFine:
          type: boolean
          description: Indica se è lo stato di fine del workflow
        accessoCondizionato:
          type: boolean
          description: Indica se le transizioni verso questo stato hanno bisogno di una conferma per essere effettuati
        nome:
          type: string
          description: Indica il nome del pulsante di transizione
        configurazioneStato:
          $ref: "#/components/schemas/ConfigurazioneStato"
...
  responses:
...
    400BadRequest:
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
      description: Bad Request
    409Conflict:
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
      description: Conflict
    404NotFound:
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
      description: Not Found
      headers:
        X-RateLimit-Limit:
          $ref: "#/components/headers/X-RateLimit-Limit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/X-RateLimit-Remaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/X-RateLimit-Reset"
...
    503ServiceUnavailable:
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
      description: Service Unavailable
      headers:
        Retry-After:
          $ref: "#/components/headers/Retry-After"
    401Unauthorized:
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
      description: Unauthorized
    default:
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
      description: Unexpected error
  headers:
...
    X-RateLimit-Limit:
      description: The number of allowed requests in the current period
      schema:
        format: int32
        type: integer
    X-RateLimit-Remaining:
      description: The number of remaining requests in the current period
      schema:
        format: int32
        type: integer
    X-RateLimit-Reset:
      description: The number of seconds left in the current period
      schema:
        format: int32
        type: integer
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer